@@ -12,49 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.4.6",
-    "@angular/cdk": "2.0.0-beta.12",
-    "@angular/common": "4.4.6",
-    "@angular/compiler": "4.4.6",
-    "@angular/core": "4.4.6",
-    "@angular/forms": "4.4.6",
-    "@angular/http": "4.4.6",
-    "@angular/material": "2.0.0-beta.12",
-    "@angular/platform-browser": "4.4.6",
-    "@angular/platform-browser-dynamic": "4.4.6",
-    "@angular/router": "4.4.6",
+    "@angular/animations": "~9.0.4",
+    "@angular/cdk": "^9.1.0",
+    "@angular/common": "~9.0.4",
+    "@angular/compiler": "~9.0.4",
+    "@angular/core": "~9.0.4",
+    "@angular/forms": "~9.0.4",
+    "@angular/material": "^9.1.0",
+    "@angular/platform-browser": "~9.0.4",
+    "@angular/platform-browser-dynamic": "~9.0.4",
+    "@angular/platform-server": "~9.0.4",
+    "@angular/router": "~9.0.4",
     "angular-material-prefix-updater": "0.0.6",
-    "class-transformer": "0.1.8",
-    "core-js": "2.5.3",
-    "konva": "1.6.3",
+    "class-transformer": "^0.2.3",
+    "core-js": "^2.6.11",
+    "konva": "^4.1.6",
     "later": "github:RickCarlino/later#fc9ceb582728f75edbf9581e4f53da5fbb8cd4ac",
-    "lodash": "4.17.4",
-    "moment": "2.20.1",
-    "ng-click-outside": "2.4.1",
-    "ng2-ace-editor": "0.2.5",
-    "ngx-clipboard": "5.1.0",
+    "lodash": "^4.17.15",
+    "moment": "^2.24.0",
+    "ng-click-outside": "^6.0.0",
+    "ng2-ace-editor": "0.3.4",
+    "ngx-clipboard": "^13.0.0",
     "prettycron": "0.10.0",
-    "rxjs": "5.5.6",
-    "zone.js": "0.8.20"
+    "rxjs": "~6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^4.4.3",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
-    "copy-webpack-plugin": "4.3.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "@angular-devkit/build-angular": "~0.900.4",
+    "@angular/cli": "~9.0.4",
+    "@angular/compiler-cli": "~9.0.4",
+    "@angular/language-service": "~9.0.4",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^13.7.7",
+    "codelyzer": "^5.1.2",
+    "copy-webpack-plugin": "^5.1.1",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.4.1",
+    "tslint": "~5.20.0",
+    "typescript": "~3.7.5"
   }
 }
