@@ -13,7 +13,8 @@
 # limitations under the License.
 
 import os
-
+import sys
+sys.path.insert(0, os.getcwd())
 from tests import utils
 
 
