@@ -13,10 +13,11 @@
 # limitations under the License.
 
 from google.appengine.ext import testbed
-import mock
-
 from core import models
 
+import os
+import sys
+sys.path.insert(0, os.getcwd())
 from tests import utils
 
 
