@@ -15,8 +15,6 @@
 import { Injectable } from '@angular/core';
 
 import { ApiService } from 'app/api.service';
-import { Stage } from 'app/models/stage';
-import { Pipeline } from 'app/models/pipeline';
 
 @Injectable()
 export class StagesService extends ApiService {
@@ -24,20 +22,21 @@ export class StagesService extends ApiService {
   private url = `${this.getHost()}/stages`;
 
   getStages() {
+    this.removeContentTypeHeader();
     return this.http.get(this.url, this.options)
                     .toPromise()
-                    .then(res => res.json() as Stage[])
                     .catch(this.handleError);
   }
 
   addStage(stage_data) {
+    this.addContentTypeHeader();
     return this.http.post(this.url, JSON.stringify(stage_data), this.options)
                     .toPromise()
-                    .then(res => res.json() as Stage)
                     .catch(this.handleError);
   }
 
   deleteStage(id) {
+    this.removeContentTypeHeader();
     return this.http.delete(this.getStageUrl(id))
                     .toPromise()
                     .catch(this.handleError);
@@ -52,10 +51,10 @@ export class StagesService extends ApiService {
   }
 
   getPipelinesForStage(stage) {
+    this.removeContentTypeHeader();
     return this.http.get(this.getPipelinesUrl(stage.sid), this.options)
                     .toPromise()
-                    .then(res => {
-                      const pipelines = res.json() as Pipeline[];
+                    .then((pipelines: any) => {
                       for (const pipeline of pipelines) {
                         pipeline.sid = stage.sid;
                       }
