@@ -13,11 +13,8 @@
 // limitations under the License.
 
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions, URLSearchParams } from '@angular/http';
 import { serialize } from 'class-transformer';
-import 'rxjs/add/operator/toPromise';
 
-import { Param } from 'app/models/param';
 import { ApiService } from 'app/api.service';
 
 @Injectable()
@@ -26,47 +23,45 @@ export class JobsService extends ApiService {
   private url = `${this.getHost()}/jobs`;
 
   getJobsByPipeline(pipeline_id) {
-    const params = new URLSearchParams();
-    params.set('pipeline_id', pipeline_id);
-    this.options.search = params;
+    this.options.params = {'pipeline_id': pipeline_id};
+    this.removeContentTypeHeader();
     return this.http.get(this.url, this.options)
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
   getJob(id) {
+    this.removeContentTypeHeader();
     return this.http.get(this.getJobUrl(id))
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
   addJob(job) {
-    return this.http.post(this.url, serialize(job), { headers: this.headers })
+    this.addContentTypeHeader();
+    return this.http.post(this.url, serialize(job), { headers: this.options.headers })
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
   updateJob(job) {
-    return this.http.put(this.getJobUrl(job.id), serialize(job), { headers: this.headers })
+    this.addContentTypeHeader();
+    return this.http.put(this.getJobUrl(job.id), serialize(job), { headers: this.options.headers })
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
   deleteJob(id) {
+    this.removeContentTypeHeader();
     return this.http.delete(this.getJobUrl(id))
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
   startJob(id) {
-    return this.http.post(this.getJobUrl(id) + '/start', null, { headers: this.headers })
+    this.addContentTypeHeader();
+    return this.http.post(this.getJobUrl(id) + '/start', null, { headers: this.options.headers })
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
