@@ -14,9 +14,7 @@
 
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { PipelineParamsComponent } from './pipeline-params.component';
 
 describe('PipelineParamsComponent', () => {
@@ -25,6 +23,10 @@ describe('PipelineParamsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [
+        FormsModule,
+        ReactiveFormsModule
+      ],
       declarations: [ PipelineParamsComponent ]
     })
     .compileComponents();
