@@ -14,9 +14,7 @@
 
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-
+import { RouterModule } from '@angular/router';
 import { PipelineGraphComponent } from './pipeline-graph.component';
 
 describe('PipelineGraphComponent', () => {
@@ -25,6 +23,7 @@ describe('PipelineGraphComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [ RouterModule.forRoot([]) ],
       declarations: [ PipelineGraphComponent ]
     })
     .compileComponents();
