@@ -13,11 +13,8 @@
 // limitations under the License.
 
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions, URLSearchParams } from '@angular/http';
-import 'rxjs/add/operator/toPromise';
 
 import { ApiService } from 'app/api.service';
-import { Pipeline } from 'app/models/pipeline';
 
 @Injectable()
 export class PipelinesService extends ApiService {
@@ -25,74 +22,74 @@ export class PipelinesService extends ApiService {
   private url = `${this.getHost()}/pipelines`;
 
   getPipelines() {
+    this.removeContentTypeHeader();
     return this.http.get(this.url, this.options)
                     .toPromise()
-                    .then(res => res.json() as Pipeline[])
                     .catch(this.handleError);
   }
 
   getPipeline(id) {
+    this.removeContentTypeHeader();
     return this.http.get(this.getPipelineUrl(id))
                     .toPromise()
-                    .then(res => res.json() as Pipeline)
                     .catch(this.handleError);
   }
 
   addPipeline(pipeline) {
+    this.addContentTypeHeader();
     return this.http.post(this.url, JSON.stringify(pipeline), this.options)
                     .toPromise()
-                    .then(res => res.json() as Pipeline)
                     .catch(this.handleError);
   }
 
   updatePipeline(pipeline) {
+    this.addContentTypeHeader();
     return this.http.put(this.getPipelineUrl(pipeline.id), JSON.stringify(pipeline), this.options)
                     .toPromise()
-                    .then(res => res.json() as Pipeline)
                     .catch(this.handleError);
   }
 
   deletePipeline(id) {
+    this.removeContentTypeHeader();
     return this.http.delete(this.getPipelineUrl(id))
                     .toPromise()
                     .catch(this.handleError);
   }
 
   startPipeline(id) {
+    this.addContentTypeHeader();
     return this.http.post(this.getPipelineUrl(id) + '/start', {}, this.options)
                     .toPromise()
-                    .then(res => res.json() as Pipeline)
                     .catch(this.handleError);
   }
 
   stopPipeline(id) {
+    this.addContentTypeHeader();
     return this.http.post(this.getPipelineUrl(id) + '/stop', {}, this.options)
                     .toPromise()
-                    .then(res => res.json() as Pipeline)
                     .catch(this.handleError);
   }
 
   importPipeline(file: File) {
     const formData: FormData = new FormData();
     formData.append('upload_file', file, file.name);
-    const headers = new Headers();
-    const options = new RequestOptions({ headers: headers });
-    return this.http.post(this.url + '/import', formData, options)
+    this.removeContentTypeHeader();
+    return this.http.post(this.url + '/import', formData)
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
   exportPipeline(id) {
-    return this.http.get(this.getPipelineUrl(id) + '/export')
+    this.removeContentTypeHeader();
+    return this.http.get(this.getPipelineUrl(id) + '/export', {observe: 'response'})
                     .toPromise()
                     .catch(this.handleError);
   }
 
   updateRunOnSchedule(id, run_on_schedule) {
+    this.addContentTypeHeader();
     return this.http.patch(this.getPipelineUrl(id) + '/run_on_schedule', {run_on_schedule: run_on_schedule}, this.options)
                     .toPromise()
-                    .then(res => res.json() as Pipeline)
                     .catch(this.handleError);
   }
 
@@ -101,18 +98,18 @@ export class PipelinesService extends ApiService {
   }
 
   getLogs(id, params) {
-    const url = this.getPipelineUrl(id) + '/logs';
-    const p = new URLSearchParams();
+    const url = this.getPipelineUrl(id) + '/logs'
+    const p = {};
     for (const k of Object.keys(params)) {
       if (params[k] !== null) {
-        p.set(k, params[k]);
+        p[k] = params[k];
       }
     }
 
-    this.options.search = p;
+    this.options.params = p;
+    this.removeContentTypeHeader();
     return this.http.get(url, this.options)
                     .toPromise()
-                    .then(res => res.json())
                     .catch(this.handleError);
   }
 
