@@ -13,7 +13,8 @@
 // limitations under the License.
 
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { HttpClientModule } from '@angular/common/http';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { SettingsComponent } from './settings.component';
 
 describe('SettingsComponent', () => {
@@ -22,6 +23,11 @@ describe('SettingsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [
+        FormsModule,
+        HttpClientModule,
+        ReactiveFormsModule
+      ],
       declarations: [ SettingsComponent ]
     })
     .compileComponents();
